feat(allocation-strategy): 新增RATIO数据分配策略支持按比例分发数据
新增RATIO枚举项，支持基于预定义比例的数据分配机制。每个VM根据
配置的比例权重（如7:2:1）接收相应数量的数据样本，适用于算力不
均的联邦学习场景。

feat(task-dto): 调整dataRatio为千分比权重并增强校验规则

将TaskCreateDTO和TaskParticipant中的dataRatio字段由Double类型改为
Integer类型，取值范围为1-1000，表示数据分配的千分比权重。所有参
与者的权重总和必须等于1000，支持更精确的数据分配控制。

feat(workflow): 增强初始化阶段上下文信息设置

在初始化阶段处理器中增加从数据库读取任务信息、设置数据集ID、分
发策略及目标VM列表到工作流上下文的功能，为后续数据分发阶段提供
必要参数支持。

feat(data-distribution): 新增数据切片增强功能接口

扩展DataDistributionService接口，新增支持数据切片分发、批次范围
计算及双重索引添加的方法，用于实现更精细化的数据分发与管理。

feat(websocket-protocol): 支持数据完整性验证及增强响应内容

在WebSocket协议服务中集成数据分发与切片验证服务，支持解析并验证
数据集完成消息中的SliceVerification信息，并在响应中返回详细的验
证结果与缺失数据指标。

+++ backend-springboot/feduwacomm-pojo/src/main/java/com/feduwacomm/entity/TaskParticipant.java
@@ -39,7 +39,23 @@ public class TaskParticipant {
 
     // v1.3 新增字段
     private String participantId;
-    private Double dataRatio;
+    /**
+     * v1.5.1.1 数据分配千分比权重
+     *
+     * <p>取值范围：1-1000的正整数
+     * <p>约束条件：同一联邦学习任务的所有参与者的dataRatio之和必须等于1000
+     * <p>语义说明：dataRatio表示该VM占用数据集的千分比
+     *
+     * <p>示例：
+     * <ul>
+     *   <li>VM1.dataRatio=700, VM2.dataRatio=200, VM3.dataRatio=100 (7:2:1比例)</li>
+     *   <li>VM1占用70%数据(700/1000), VM2占用20%(200/1000), VM3占用10%(100/1000)</li>
+     *   <li>总和验证: 700+200+100=1000 ✓</li>
+     * </ul>
+     *
+     * <p>如果为null，系统会根据参与者数量自动分配平均权重（如3个VM各334、333、333）
+     */
+    private Integer dataRatio;
     private String capabilities; // JSON格式存储能力列表
     private Integer maxCpuUsage;
     private Integer maxMemoryUsage;
